some tweaks to reduce picky compiler warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@3812 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/Unarchive.cpp
@@ -266,7 +266,7 @@ int ExpandZip(const char *archive_fname, const char *prepend_path)
 		{
 //			if (zip.GetFileInfo(fh, i))
 //				printf("%d: %s\n", i, fh.GetFileName());
-			zip.ExtractFile(i, path.c_str());
+			zip.ExtractFile((WORD)i, path.c_str());
 		}
 	}
 	catch(...)
